peewee-ify! site working, scripts not; bingo and csv not portable yet
