--- debug.cfg
@@ -1,1 +1,0 @@
-DEBUG = True
