+++ ptv_helper/config/default.py
@@ -1,0 +1,8 @@
+import os
+
+DATABASE = os.path.abspath(
+    os.path.join(os.path.dirname(__file__), '../../ptv.db'))
+SECRET_KEY = '9Zbl48DxpawebuOKcTIxsIo7rZhgw2U5qs2mcE5Hqxaa7GautgOh3rkvTabKp'
+ADMINS = ['dougal@gmail.com']
+
+DEBUG = True
