+++ ptv_helper/config/deploy_base.py
@@ -1,0 +1,14 @@
+import logging
+from logging.handlers import SMTPHandler
+import sys
+
+from .default import *
+
+DEBUG = False
+
+
+logging.basicConfig(stream=sys.stderr)
+mail_handler = SMTPHandler('127.0.0.1', 'helper@previously.tv', ADMINS,
+                           "[ptv-helper] blew up")
+mail_handler.setLevel(logging.ERROR)
+app.logger.addHandler(mail_handler)
