+++ ptv_helper/static/ZeroClipboard.swf
