+++ ptv_helper/templates/login.html
@@ -34,7 +34,7 @@
       <option selected disabled value='-'>-</option>
       <option id='pick-someone-else' value=''>Someone New and Special</option>
       {% for mod in mods %}
-        <option value="{{ mod['id'] }}">{{ mod['name'] }}</option>
+        <option value="{{ mod.id }}">{{ mod.name }}</option>
       {% endfor %}
     </select>
