+++ ptv_helper/templates/my_shows_next.html
@@ -36,14 +36,14 @@
         <tr>
           <td><a href="{{ show_forum_id|forum_url }}">{{ showname }}</a></td> <td>
             {% if last %}
-              <span class="date">{{ last['first_aired'] or '?' }}</span>
-              S{{ last['season_number'] }}
-              E{{ last['episode_number'] }}:
-              {{ last['name'] or '-' }}
-              (<a href="http://thetvdb.com/?tab=episode&seriesid={{ last['seriesid'] }}&seasonid={{ last['seasonid'] }}&id={{ last['episodeid'] }}">tvdb</a>)
-              {% if last['overview'] %}
+              <span class="date">{{ last.first_aired or '?' }}</span>
+              S{{ last.season_number }}
+              E{{ last.episode_number }}:
+              {{ last.name or '-' }}
+              (<a href="http://thetvdb.com/?tab=episode&amp;seriesid={{ last.seriesid }}&amp;seasonid={{ last.seasonid }}&amp;id={{ last.episodeid }}">tvdb</a>)
+              {% if last.overview %}
                 (<a class="clip" title="Click to copy." href="#"
-                    data-clipboard-text="[quote]{{ last['overview'] }}[/quote]">copy desc</a>)
+                    data-clipboard-text="[quote]{{ last.overview }}[/quote]">copy desc</a>)
               {% endif %}
 
             {% else %}
@@ -52,14 +52,14 @@
           </td>
           <td>
             {% if next %}
-              <span class="date">{{ next['first_aired'] or '?' }}</span>
-              S{{ next['season_number'] }}
-              E{{ next['episode_number'] }}:
-              {{ next['name'] or '-' }}
-              (<a href="http://thetvdb.com/?tab=episode&seriesid={{ next['seriesid'] }}&seasonid={{ next['seasonid'] }}&id={{ next['episodeid'] }}">tvdb</a>)
-              {% if next['overview'] %}
+              <span class="date">{{ next.first_aired or '?' }}</span>
+              S{{ next.season_number }}
+              E{{ next.episode_number }}:
+              {{ next.name or '-' }}
+              (<a href="http://thetvdb.com/?tab=episode&amp;seriesid={{ next.seriesid }}&amp;seasonid={{ next.seasonid }}&amp;id={{ next.episodeid }}">tvdb</a>)
+              {% if next.overview %}
                 (<a class="clip" title="Click to copy." href="#"
-                    data-clipboard-text="[quote]{{ next['overview'] }}[/quote]">copy desc</a>)
+                    data-clipboard-text="[quote]{{ next.overview }}[/quote]">copy desc</a>)
               {% endif %}
             {% else %}
               &mdash;
