+++ ptv_helper/templates/turf_row.html
@@ -1,15 +1,15 @@
-<tr id="show-{{ show['id'] }}" data-showid="{{ show['id'] }}"
-    class="{% if info['n_posts'] != 'n/a' and info['n_posts'] > hi_post_thresh %}high-posts{% endif -%}
+<tr id="show-{{ show.id }}" data-showid="{{ show.id }}"
+    class="{% if show.n_posts() != 'n/a' and show.n_posts() > hi_post_thresh %}high-posts{% endif -%}
            {% if info['n_mods'] == 0 %} no-mods{% else %} has-mods{% endif -%}
            {% if user.is_authenticated() and info['my_info'][0] %} mine{% else %} not-mine{% endif -%}
            {% if user.is_authenticated() and info['my_info'][0] == 'g' %} my-lead{% else %} not-my-lead{% endif -%}
-           {% if show['tvdb_ids'] %} has-tvdb{% else %} no-tvdb{% endif -%}
-           {% if show['gone_forever'] %} over{% else %} not-over{% endif -%}
-           {% if show['we_do_ep_posts'] %} do-ep-posts{% else %} no-ep-posts{% endif -%}
-           {% if show['needs_leads'] %} needs-leads{% else %} no-needs-leads{% endif -%}
-           {% if show['needs_backups'] %} needs-backups{% else %} no-needs-backups{% endif %} {{ parity }}">
+           {% if show.tvdb_ids %} has-tvdb{% else %} no-tvdb{% endif -%}
+           {% if show.gone_forever %} over{% else %} not-over{% endif -%}
+           {% if show.we_do_ep_posts %} do-ep-posts{% else %} no-ep-posts{% endif -%}
+           {% if show.needs_leads %} needs-leads{% else %} no-needs-leads{% endif -%}
+           {% if show.needs_backups %} needs-backups{% else %} no-needs-backups{% endif %} {{ parity }}">
   <td class="show-name">
-    <a href="{{ show['forum_id']|forum_url }}">{{ show['name'] }}</a>
+    <a href="{{ show.forum_id|forum_url }}">{{ show.name }}</a>
 
     <span class="tvdb-link">
       ({{ show.tvdb_ids | tvdb_links | safe }})
@@ -17,47 +17,47 @@
   </td>
 
   <td class="activity">
-    {{ info['n_posts']|int|commify }}
+    {{ show.n_posts()|int|commify }}
   </td>
 
   <td class="ep-threads">
     <label>
     Over:
-    <input type="checkbox" class="in-canceled" id="canceled-{{ show['id'] }}"
-           {% if show['gone_forever'] %}checked{% endif %} />
+    <input type="checkbox" class="in-canceled" id="canceled-{{ show.id }}"
+           {% if show.gone_forever %}checked{% endif %} />
     </label>
     <br />
     <label>
     Per-ep:
-    <input type="checkbox" class="in-ep-posts" id="ep-posts-{{ show['id'] }}"
-           {% if show['we_do_ep_posts'] %}checked{% endif %} />
+    <input type="checkbox" class="in-ep-posts" id="ep-posts-{{ show.id }}"
+           {% if show.we_do_ep_posts %}checked{% endif %} />
     </label>
   </td>
 
   <td class="needs">
     <label>
       Leads
-      <input type="checkbox" class="in-needs-leads" id="needs-leads-{{ show['id'] }}"
-             {% if show['needs_leads'] %}checked{% endif %} />
+      <input type="checkbox" class="in-needs-leads" id="needs-leads-{{ show.id }}"
+             {% if show.needs_leads %}checked{% endif %} />
     </label>
     <br />
     <label>
       Help
-      <input type="checkbox" class="in-needs-backups" id="needs-backups-{{ show['id'] }}"
-             {% if show['needs_backups'] %}checked{% endif %} />
+      <input type="checkbox" class="in-needs-backups" id="needs-backups-{{ show.id }}"
+             {% if show.needs_backups %}checked{% endif %} />
     </label>
   </td>
 
   {% if user.is_authenticated() %}
   {% set mystatus, mycomments = info['my_info'] %}
   <td class="you">
-    <select class="havingness" id="havingness-{{ show['id'] }}" autocomplete="off">
+    <select class="havingness" id="havingness-{{ show.id }}" autocomplete="off">
       <option value="">-</option>
       <option value="g" {% if mystatus == "g" %}selected{% endif %}>lead</option>
       <option value="c" {% if mystatus == "c" %}selected{% endif %}>helper</option>
       <option value="w" {% if mystatus == "w" %}selected{% endif %}>could help</option>
     </select>
-    <input type="text" class="comments" id="comments-{{ show['id'] }}"
+    <input type="text" class="comments" id="comments-{{ show.id }}"
            placeholder="Comments" value="{{ mycomments or "" }}" size="20" />
   </td>
   {% endif %}
