+++ requirements.txt
@@ -2,11 +2,12 @@ Flask==0.10.1
 Flask-Login==0.2.11
 Jinja2==2.7.3
 MarkupSafe==0.23
-Werkzeug==0.9.6
+Werkzeug==0.10.4
 cssselect==0.9.1
 itsdangerous==0.24
 lxml==3.3.5
 mechanize==0.2.5
+peewee==2.6.1
 tvdb-api==1.9
 unicodecsv==0.9.4
 wsgiref==0.1.2
